Move SettingsModal to separate component
